LayoutTests/fast/forms/percent-height-auto-width-form-controls.html

<!DOCTYPE html>
<p>The form controls of the same type below should be the same width.</p>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <input id="text-a" style="height: 100%">
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <input id="text-b">
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <input id="file-a" type=file style="height: 100%">
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <input id="file-b" type=file>
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <input id="range-a" type=range style="height: 100%">
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <input id="range-b" type=range>
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <select id="select-a" style="height: 100%"><option>aaaaaaaaaaaaaaaaaaaaaaaaa
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <select id="select-b"><option>aaaaaaaaaaaaaaaaaaaaaaaaa
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <select id="multiple-a" multiple style="height: 100%"><option>aaaaaaaaaaaaaaaaaaaaaaaaa
    </td></tr></tbody></table>
</div>

<div style="width: 100px; background: pink">
    <table><tbody><tr><td>
        <select id="multiple-b" multiple><option>aaaaaaaaaaaaaaaaaaaaaaaaa
    </td></tr></tbody></table>
</div>

<script src="../js/resources/js-test-pre.js"></script>
<script>
var ids = ['text', 'file', 'range', 'select', 'multiple'];
ids.forEach(function(id) {
    shouldBe('document.getElementById("' + id + '-a").offsetWidth', 'document.getElementById("' + id + '-b").offsetWidth');
})
</script>
<script src="../js/resources/js-test-post.js"></script>
