third_party/blink/web_tests/fast/forms/percent-height-auto-width-form-controls-expected.txt

PASS document.getElementById("text-a").offsetWidth is document.getElementById("text-b").offsetWidth
PASS document.getElementById("file-a").offsetWidth is document.getElementById("file-b").offsetWidth
PASS document.getElementById("range-a").offsetWidth is document.getElementById("range-b").offsetWidth
PASS document.getElementById("select-a").offsetWidth is document.getElementById("select-b").offsetWidth
PASS document.getElementById("multiple-a").offsetWidth is document.getElementById("multiple-b").offsetWidth
PASS successfullyParsed is true

TEST COMPLETE


The form controls of the same type below should be the same width.

aaaaaaaaaaaaaaaaaaaaaaaaa
aaaaaaaaaaaaaaaaaaaaaaaaa
aaaaaaaaaaaaaaaaaaaaaaaaa
aaaaaaaaaaaaaaaaaaaaaaaaa
